Remember latest post time per feed, not latest run time
FastMail's blog published a post around 8am my time with a "published" time of 1pm. This screwed with the bot, because it constantly thought it was a new post. This change makes the assumption that new posts will always have a publish time greater than all previous posts. This assumption may not always hold true, but it's a stronger assumption than the assumption that publish time represents the time a piece of content is made available.